Now it's simple to add crud funcs for any model by use crud func for router (and provide some schemas for it)
